fix(blocks): make the mint-audit emitter structurally non-throwing via a narrow field type
`emitMintAuditToStdout` called `JSON.stringify` on `Record<string, unknown>` inside
a token-mint request handler. None of the five callers can trigger it — they pass
only strings, numbers, booleans and string[] — so this was not a live bug. But the
`unknown` value type meant a future caller could make an audit log throw and turn a
successful mint into a 500. An audit log must never be able to break the operation
it audits.

Fixed structurally rather than with a try/catch. A `catch` would trade a crash for
a silently missing audit line, which is the same blind spot this module exists to
close; an unrepresentable value is a compile error at the call site instead.

  type JsonScalar = string | number | boolean | null;
  export type MintAuditFields =
    Record<string, JsonScalar | readonly JsonScalar[] | undefined>;

`JSON.stringify` has exactly two throwing inputs — a BigInt and a circular
structure — and both are now unrepresentable: a BigInt is not a `JsonScalar`, and a
cycle needs a nested object or array reference, which `readonly JsonScalar[]`
cannot hold. Excluding nested objects also rules out a throwing custom `toJSON`.
`undefined` stays permitted so a caller can omit a key, which is what preserves the
three-valued `requestBudgetedSpend` signal the #3703 step-3 gate reads.

All five production call sites type-check UNCHANGED — no call site needed a
widening. `app.ephemeralSource` is `'pending' | 'brand-new'` (optional) and
`app.status` is `string`, both string-like.

The existing three-valued signal tests (absent stays absent, false stays false,
true stays true) pass unchanged under the narrowed type and are not duplicated.

Adds compile-time guards asserting a BigInt and a nested object are rejected, via
`@ts-expect-error` — which fails typecheck if the error does NOT occur, so widening
the type back to `unknown` turns them red instead of leaving decoration behind.

Two things had to be fixed to make those guards real, both caught by running the
mutation rather than by reading the code:

  - MOVED the unit test out of `__tests__/`. `tsconfig.json` excludes
    `src/**/__tests__/**`, so the file was never typechecked and the directives were
    inert: widening the type produced zero new errors. It now sits directly in
    `src/server/logging/`, matching the sibling `central-error-log.test.ts` /
    `whatif-error-logging.test.ts` convention, and is in tsc's program.

  - BUILD the bigint with `BigInt(1)`, not the `1n` literal. This repo targets
    ES2018, so `1n` is itself a target error (TS2737); an `@ts-expect-error` over it
    was satisfied by THAT and stayed green with the field type widened — green for
    the wrong reason. With `BigInt(1)` the only error is the assignability one.

Verified: widening `MintAuditFields` back to `Record<string, unknown>` now yields
14 typecheck errors (12 pre-existing + both directives reported unused); the narrow
type yields 12, matching base exactly.
